+++ src/PowerModelsDistribution.jl
@@ -27,6 +27,7 @@ module PowerModelsDistribution
     include("form/acr.jl")
     include("form/apo.jl")
     include("form/dcp.jl")
+    include("form/ivr.jl")
     include("form/bf.jl")
     include("form/bf_mx.jl")
     include("form/bf_mx_lin.jl")
@@ -46,13 +47,13 @@ module PowerModelsDistribution
 
     include("prob/mld.jl")
     include("prob/opf.jl")
-    include("prob/opf_lm.jl")
+    include("prob/opf_iv.jl")
     include("prob/opf_oltc.jl")
     include("prob/opf_bf.jl")
     include("prob/opf_bf_lm.jl")
     include("prob/pf.jl")
-    include("prob/pf_lm.jl")
     include("prob/pf_bf.jl")
+    include("prob/pf_iv.jl")
     include("prob/debug.jl")
     include("prob/test.jl")