+++ src/prob/opf_iv.jl
@@ -1,0 +1,59 @@
+""
+function run_mc_opf_iv(data::Dict{String,Any}, model_type, solver; kwargs...)
+    return _PMs.run_model(data, model_type, solver, build_mc_opf_iv; ref_extensions=[ref_add_arcs_trans!], multiconductor=true, kwargs...)
+end
+
+
+""
+function run_mc_opf_iv(file::String, model_type, solver; kwargs...)
+    return run_mc_opf_iv(PowerModelsDistribution.parse_file(file), model_type, solver; kwargs...)
+end
+
+
+""
+function build_mc_opf_iv(pm::_PMs.AbstractPowerModel)
+    # Variables
+    variable_mc_voltage(pm)
+    variable_mc_branch_current(pm)
+    variable_mc_transformer_current(pm)
+    variable_mc_generation(pm)
+    variable_mc_load(pm)
+
+    # Constraints
+    for i in _PMs.ids(pm, :ref_buses)
+        constraint_mc_theta_ref(pm, i)
+    end
+
+    # gens should be constrained before KCL, or Pd/Qd undefined
+    for id in _PMs.ids(pm, :gen)
+        constraint_mc_generation(pm, id)
+    end
+
+    # loads should be constrained before KCL, or Pd/Qd undefined
+    for id in _PMs.ids(pm, :load)
+        constraint_mc_load(pm, id)
+    end
+
+    for i in _PMs.ids(pm, :bus)
+        constraint_mc_current_balance_load(pm, i)
+    end
+
+    for i in _PMs.ids(pm, :branch)
+        constraint_mc_current_from(pm, i)
+        constraint_mc_current_to(pm, i)
+
+        constraint_mc_voltage_drop(pm, i)
+
+        constraint_mc_voltage_angle_difference(pm, i)
+
+        constraint_mc_thermal_limit_from(pm, i)
+        constraint_mc_thermal_limit_to(pm, i)
+    end
+
+    for i in _PMs.ids(pm, :transformer)
+        constraint_mc_trans(pm, i)
+    end
+
+    # Objective
+    _PMs.objective_min_fuel_cost(pm)
+end