+++ test/pf.jl
@@ -157,14 +157,14 @@
         end
 
         @testset "matrix branch shunts acp pf" begin
-            sol = PMD.run_ac_mc_pf_lm("../test/data/opendss/case_mxshunt.dss", ipopt_solver)
+            sol = PMD.run_ac_mc_pf("../test/data/opendss/case_mxshunt.dss", ipopt_solver)
 
             @test all(isapprox.(sol["solution"]["bus"]["2"]["vm"], [0.987399, 0.981300, 1.003536]; atol=1E-6))
         end
 
         @testset "matrix branch shunts acr pf" begin
             data_pmd = PMD.parse_file("../test/data/opendss/case_mxshunt.dss")
-            pm = PMs.instantiate_model(data_pmd, PMs.ACRPowerModel, PMD.build_mc_pf_lm, ref_extensions=[PMD.ref_add_arcs_trans!], multiconductor=true)
+            pm = PMs.instantiate_model(data_pmd, PMs.ACRPowerModel, PMD.build_mc_pf, ref_extensions=[PMD.ref_add_arcs_trans!], multiconductor=true)
             sol = PMs.optimize_model!(pm, optimizer=ipopt_solver)
 
             calc_vm(id) = sqrt.(sol["solution"]["bus"][id]["vr"].^2+sol["solution"]["bus"][id]["vi"].^2)